Properly match the LXC index item to download
Currently the match will return two results if the
variant to download is 'default' and the index
contains both 'default' and 'default-beta.1'. The
download task then downloads them both and the last
one wins.

This patch corrects the matching mechanism to ensure
that the match is properly delimited to prevent
improper matches.

This patch also ensures that only one download is done
if there are multiple matches. The list of results is
reverse sorted so that the most recent result is used.

Change-Id: I1086d41bd483d91d8d037a3ba8db4f92e738536e