Add transaction cleanup to cent prep
Because we use a transation when prep'ing the cent image we need to
cleanup the transactions upon completion. This change is looking to
solve the issue where ansible will fail package install reporting

> "The following packages have pending transactions"

This seems to be happening totally randomly though is always early in
the deployment process. After spot checking a few instances I
noticed the command found several stale transactions and cleaned them
up. Example from within a contianer after a failed build.

> http://paste.openstack.org/show/664394/

Since incorporating this change I've not seen the issue.

Change-Id: I618c767c6daace11cf12d810d7bef47eec15f240
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>