Clear the cached machine-id
The systemd machine-id needs to be unique, at present our cached ID is
being stored which is then being cloned upon container create. This
change simply removes the machine-id files from the cache so that we're
not re-using ID's throughout the cloud.

More information on the systemd machine-id can be found here:
https://www.freedesktop.org/software/systemd/man/machine-id.html

Change-Id: Ie20e3f01bf818f7c5510d6bffd19abdce8b38180
Depends-On: https://review.openstack.org/#/c/544375
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>