Maintain backwards compatibility for the machinectl volume size
In I85fefc6ce186bb6808ac37a9ea79a50e29671115 the format for
lxc_host_machine_volume_size was changed from a forced GB
value to a value where the size can be specified by the user.

Unfortunately this totally breaks previous settings, as
demonstrated by the need to implement this patch to get the
integrated build to work again:
https://review.openstack.org/544724

Unfortunately this means that any value set in any previous
series will also fail, meaning that an upgrade from a previous
setting in a previous or same series will fail.

This is not an acceptable outcome, so this patch forces it
back to using Gigabytes as documented and makes sure it's
consistent across the old/new systemd implementations. It also
converts any other values set (perhaps with trailing G/K) to
use G so that regardles sof someone setting it in the interim
state it will still work.

Change-Id: Iee220190b92613707f890929aad855dec46892da