Install common packages into container cache
The pip_install role adds several packages to each container and this
is a time consuming process since it runs in each container. This
patch copies those packages into the lxc_hosts role so that each
package is installed only one time into the cache.

This should speed up deployments and testing.

Change-Id: I7793d76c01ff77d963b92e54c762426cfdf94e94