Implement always tag for required fact
The fact set for 'Set image index fact' is used
in the next task, which is tagged 'always' but
the fact setting is not. This breaks when using
tags.

Given that all the tasks in this file are set to
'always', this patch moves the tag setting to
the inclusion instead.

Change-Id: If319e749ef7cdfbbbd3d03096608fc736927f89d