templates: lxc-system-manage: Send correct DNS over dhcp
We should set the correct DNS IP address via DHCP so containers can get
a valid entry. This fixes a problem where the DNS inherits the DNS
server from the host and the host uses the localhost IP in case it's
running a local DNS caching resolver such as dnsmasq.

Depends-On: Ied7632037f737c3f32c34dac70531065c54496c9
Change-Id: I14f8373897da28dea2ea03500c2be46c5b40d51c