tasks: Ensure machinectl quota value is reasonable
It appears that when setting a limit which is greater than the size
of the underlying filesystem, machinectl does not change the quota at
all. As such, lets calculate the limit based on the /var/lib/machines
mountpoint or any of its parents.

Change-Id: I8ddc9931cbca5db8c1a411fd2581b34763179d01
(cherry picked from commit 345b501464880ecbb5d122d7715277cf1d0506c5)