Remove generic default interfaces
With the implementation of networkd the ENI scripts and config files for
the default interfaces shipped with the lxc container images we use is
no longer useful. These old files can cause conflicts in networking
should the old scripts and networkd get confused especially when it
comes to an interface that is setup for DHCP. This change simply removes
the default interfaces for cent, suse, and ubuntu are removed.

Change-Id: I59fa50216fca320be6643585159fcb72c1925a2f
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>