Fix lxc cache prep resolvers
Recent changes in the upstream LXC images have resulted in breakage
caused by busted symlinks, likely to the image moving to
"systemd-resolved". This change backports the process we use in master
to create the cache resolvers which will ensure we're not having to deal
with this issue.

Change-Id: I3769053956b958a0ac7cab499ea90af2a40de850
Related-Change: I1775e775bbb7fe86bdffdd8296c2cff5ebc5bac8
Closes-Bug: #1768592
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>