Fix LXC cache resolvers more appropriately
The implementation of I3769053956b958a0ac7cab499ea90af2a40de850
was somewhat brutal in the approach based on the assumption that
systemd networking is implemented in the base cache. This may be
true for Rocky, but is not true for any of the older branches.

This patch returns the previous method, but fixes it in two ways:

1. It uses the -f file test operator, instead of the deprecated
   -a operator.
2. It also tests for the presence of a symbolic link with -L.

The downloaded image has resolvconf installed, so there is a symlink
for /etc/resolv.conf which points to /run/resolvconf/resolv.conf
in the cache. Given that in a chroot /run does not exist, the file
test operator fails. This is why we're adding the symlink check.

Instead of then removing resolv.conf at the end, we return it back
so that resolveconf works as expected. This is important for anyone
implementing their resolvers through configuration of network
interfaces.

This is only implemented for Queens and older because for master
we've moved to using a different image source, and using
systemd-resolved.

Closes-Bug: #1768592
Change-Id: I11d069e2d0cb3becad067689fa8f1e4c9391f368