Setup /dev/random and /dev/urandom device in cache prep (ubuntu/suse)
In the cache preparation stage, there are certain libraries that
depend on the existance of /dev/random and /dev/urandom in order
to be able to function correctly, such as NSS in the latest CentOS
release (7.5)

This patch adds those nodes so that the libraries are able to use
them with no problems, allowing yum and rpm specifically to work
properly again.

This was done for all distributions in [1], but the backports only
did it for RedHat.

[1] https://review.openstack.org/567585

Change-Id: Ib59aad2f2577d2f67c3ca6dc4ec20cfadd8ee960
(cherry picked from commit 741e6640c3d7d50f00d73d44c3e9c2d89ebfcd74)