SUSE: Add support for openSUSE Leap 15
This also ensures that the OBS Virtualization repository is only created
for Leap 42.X since Leap 15.X already contains the necessary LXC
packages in the standard repos. Finally, we don't use the major version
for the Leap 15 files since there is a currently a bug in Ansible[1]
which returns 'NA' as major version for Leap 15.

This patch does not add any jobs since there is a circular dependency
with the lxc_container_create role. Once Leap 15 is supported in that
role, we will add jobs in this role as well.

[1]: https://github.com/ansible/ansible/issues/41410
Change-Id: I64e38ac7b7e8dec81b5f637301df8a112aa03d9b