vars: Do not install devel packages by default
There is no good reason why we need to install devel packages in the
containers or the host. If devel packages are required to build pip
packages then the individual roles should pull them in instead. Dropping
these packages makes the containers slightly smaller and also drops the
time it takes to bootstrap them.

Change-Id: I813b4800dd3ce22f5f5d7d7cd0ac4fce5661b1e3