redhat: Add the EPEL repository for aria2 and python2-lxc packages
The ARIA2 package is needed from EPEL so we install it inside
the role. Moreover, in order to minimize the potential EPEL vs RDO
package conflicts, we only allow a small subset of packages from the
EPEL repository. This also removes the 'debootstrap' package from
CentOS since it's not needed. Finally, we include the nodepool playbook
from the tests repository in order to populate the necessary CI facts.

Depends-On: https://review.openstack.org/#/c/574801/
Depends-On: https://review.openstack.org/#/c/575059/
Change-Id: Icd4d09a26cd95ecfefec26142723ddffe81aae74
Co-Authored-By: Markos Chandras <mchandras@suse.de>