Use opensuse_mirror if it's set
To enable a common variable for overriding the OpenSUSE mirror,
we use opensuse_mirror by default and provide a default value
if it is not set. This ensures that a common variable can be used
to set them all between lxc_hosts and openstack_hosts. Doing it
this way also means that we do not need to add any 'glue' in the
group_vars to keep them consistent.

Change-Id: I22ac019d8783bc8e6c9d71c22c04314b77c55261
(cherry picked from commit d5995da9dfb1d8035e18fcc2eb73e94e926ca2e1)