tasks: lxc_apparmor: Fix links for disabled profiles
The /etc/apparmor.d/local/* files are for user overrides. Symlinking
these to disable does not disable the profile since the main one still
lives in /etc/apparmor.d/. Moreover, the symlink destination for the
'ping' profile was wrong so we fix the typo to use a unique name for
that profile.

Change-Id: Iba8095f1068bc6f00064dd31b92cea5931a33b90