Simplify inserting apt keys into LXC image
Bionic requires a functioning gpg-agent to run apt-key add. This means
that gpg-agent must be working properly in the chroot when the lxc image
preparation script runs.

Previous changes [1] have enabled apt-key to communicate with gpg-agent
during the lxc_hosts role checks, however the cache prep fails almost
every time when lxc_hosts is used within other role checks.

This is not a new issue, debian-installer is affected too [2].

This change adopts the same route as d-i, and simply copies the host
/etc/apt/trusted.gpg.d directory to the LXC image, removing the
need for apt-key and in turn gpg-agent.

[1] https://review.openstack.org/#/c/588629/
[2] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=851774

Change-Id: Ia511881f675cd88ecc58360b18531f0dac486b70