Remove the checksum fills
The checksum fill was created a long time ago in a cloud far far away.
Checksums on the lxc bridge were needed back in the days when Icehouse
was a popular band and Juno was a funny movie. These tasks are no longer
required and should be removed as they do nothing but spam DMESG with
kernel stacktraces and nothing about that is cute or funny.

Change-Id: I2242725682056491106ab8012ad1fa9d0e7f9ecd
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 7e38fbb137543aa6a98af2be202885932b116985)