Drop SELinux support for CentOS 7
We do not have a maintainer at the moment for SELinux and hopefully
we will adopt the upstream openstack-selinux package, but for now
in order to let deploys in environments where SELinux is set to
permissive work, we'll have to remove these bits.

This change can be reverted whenever we have a maintainer that's
available to do the work required.

Change-Id: I61141da3a391a99cb111733eae22cc7c54ce48c6
(cherry picked from commit a6dba10bb10d39b119af44ef7dbc13894773dd3e)