tasks: lxc_install_zypper: Use the 'state' option for external repos
The state of the Containers repository depends on the Leap version
that's being used. Before this patch, we did not override the default
value of the 'state' option so the repository was always added. However,
we don't need to use this repo for Leap 15 so we need to use the 'state'
value from the distro variables.

Change-Id: Ia41ffaa21fede271263c1757de949e65f6f30a67