Drop COPR priority settings
We no longer use priorities within CentOS and we let packages use
the natural priorities.  This task was put in place to restore to
default priority and now that in Rocky this task has been run, it
can be removed and no need to be re-ran as it will be no-op.

Change-Id: Iddf6caf674bba61151c1ed77db95efb0c9f0cf82