lxc: stop mirroring COPR locally on CentOS 7
We no longer need to mirror this locally all the time on all the
machines which includes downloading way too many packages than we
need.

We have mirorring and proxies in place in upstream to avoid these
issues and if we have any problems, we can resort to adding mirrors
or proxies in infra.

If there is network problems with these repositories, they will still
hit us if we're mirroring things.

Change-Id: Idad048de6aef569e0e3b73b09448750cf0a06b16