prep: remove old machinectl workarounds
There are a few manual workarounds that we're placing in order
to workaround old versions of machinectl however we don't actually
leverage those and they seem to be causing a dbus restart which
causes extra problems.

This patch removes those workarounds in order to prevent restarting
dbus which causes the system to start timing out on systemd-logind.

Change-Id: I86483225754a5b1c6030ef21e2c0cdf2cd908c3b
Closes-Bug: #1807405