Use pkill for lxc-dnsmasq systemd unit file
Previously psmisc was not installed by default, causing
the killall in the lxc-dnsmasq systemd unit file to
silently fail if killall wasn't found.  This prevented
running instances of dnsmasq to be torn down when
restarting the service.

This would exhibit as a problem during upgrades as
the older dnsmasq service would be running and the
new service would be unable to take over the port.
This would cause lxc containers to lose connectivity.

This commit switches to using pkill and ensures procps
is always installed by default.

Change-Id: I4fa838706f1163fd68ff68258bfc66cbf13bad94