Include libpython and rsync for centos in lxc base image
Some but not all roles use ansible-role-uwsgi to install libpython.
Those which do not require libpython to be in the container base
image.

Some roles use rsync for copying rootwrap files (cinder as example) and
in the long term this should be added to the distro packages for those
roles. In the meantime we add it back to the container base for centos
to allow other patches to merge.

Change-Id: Iaf92e78758c159db7e969a89a889c4d897f5045c