Replace linters test with integarted one
We've created integrated linters check job a while back and it's successfully
working for several releases. At the moment we experience difficulties
with future maintenance of the linters check from the openstack-ansible-tests
repo. So instead of fixing current one, we replace it with modern version of
the test.

We also drop regular functional testing as it's replaced with infra
integrated jobs.

Change-Id: I481ef8ec9df7b37b49b2adca606e4c02beee9e48