Disable apt phased updates inside containers.
Inside a chroot, phased updates are disabled [1]. This means that
the container base image always gets the latest packages regardless
of what is happening with the phasing.

At runtime, the default in Ubuntu releases 21.04 and onward is to
obey the package phasing information. This means that packages
inside the OSA built container image can be newer than the installation
candidates once the container is running, leading to installation
errors. This is particularly sensitive with source packages such as
systemd where there is a very tight version coupling between
all components leading to only one valid installation candidate.

This patch creates apt config inside the container base image to
always install the latest package version regardless of phasing.

There does not seem to be any alternative, as phasing is always
disabled during the debootstrap.

[1] https://discourse.ubuntu.com/t/phased-updates-in-apt-in-21-04/20345

Change-Id: Ia558e3aa1447220016c53349cf9dac0b822d06f4