Clean up NFV OpenvSwitch repo for container build
NFV repo is supposed to be installed using system packages, as it should
also contain nfvsigdist variable for yum.
So avoid issue with yum update at this step we drop the repo that was
copied from host.

As alternative approach, we can drop copying yum.repos.d at all, but this
can lead to an unexpected results.

Change-Id: Ia5041c7d855a9e988afc4c2a0d16fdeb6a9c357f