Do not user urljoin for deprecated script
The problem is that urljoin does drop path after FQDN which results in
wrong result URL. os.path.join on contrary will just drop first argument
So we use stupid join of the list through `/` which can add extra one
even if it's not needed. But by far it's the best thing we can do.

Change-Id: I11c83628b0c33146a36e58c9de85e4a1a99bd9ca