Delete centos-private-epel release note
This release note somehow was having the exact same UUID as the one
for repo_server repo [1] which resulted in reno never being copied
to the integrated repo and published.

While we can re-generate UUID and promote it, it's been over 6 years
since feature implementations, and updating UUID will result in reno
renderring for the current release. Thus it's more logical to delete it
reather then fix it.

[1] https://opendev.org/openstack/openstack-ansible-repo_server/src/branch/master/releasenotes/notes/centos-private-epel-3fe4c9ff68ec3a18.yaml

Change-Id: Iccd1225f39cda0f8c704215068977ec7d9c0b68a