Introduce common conf profile extension
With moving from distro-specific profiles, which are not packaged
for EL, to a common profile, we want to extend it with some parameters
which were defined by default in Debian/Ubuntu profiles.

There is no distinguishing between distros right now, as these
parameters should be still valid for EL as well.

According to the README file, /usr/share/lxc/config/common.conf.d/
directory is intended to extend the common.conf profile, which
we are using now with a custom common configuration.

It does not mean or replace any other possible custom configuration
for LXC containers we have in lxc-container-create role, but rather
intends to compensate lost parameters which were used inderectly
by loading OS-specific profiles, rather then a common one.

Change-Id: I53965eb0c978884d8a26c8300b123f23dc433e57