Revert "Remove lxc_net_mtu definition"
Original intention behind `lxc_net_mtu` is to control MTU of the
lxc net, which is not exposed to provider_networks and can't be
managed otherwise.

While this is "noop" network, it still might be useful to tune in case
you intend to have a host with L3-only network and jumbo frames.

This reverts change Icfdf4f806663f91383296501872bd059af19891c

Change-Id: I8edff72b429ef46aa4ebd6def1f97d1fbdce7a89
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>